Reorganize microservices package structure; remove unused
