+++ src/python/WMCore/MicroService/MSMonitor/MSMonitor.py
@@ -13,7 +13,7 @@ import time
 
 # WMCore modules
 from WMCore.MicroService.DataStructs.DefaultStructs import MONITOR_REPORT
-from WMCore.MicroService.Unified.MSCore import MSCore
+from WMCore.MicroService.MSCore import MSCore
 from WMCore.Services.Rucio.Rucio import Rucio
